+++ Simulation/Aivika/Stream.hs
@@ -25,7 +25,6 @@ module Simulation.Aivika.Stream
         splitStreamPrioritising,
         splitStreamFiltering,
         splitStreamFilteringQueueing,
-        splitStreamFilteringPrioritising,
         -- * Specifying Identifier
         streamUsingId,
         -- * Prefetching and Delaying Stream
@@ -398,39 +397,6 @@ splitStreamFilteringQueueing s preds x =
                 Just a  -> return a
                 Nothing -> reader pred
      return $ map (repeatProcess . reader) preds
-
--- | Split the input stream into a list of output streams
--- using the specified priorities after filtering.
---
--- The priority is updated only after successful applying the filter.
-splitStreamFilteringPrioritising :: PriorityQueueStrategy s p
-                                    => s
-                                    -- ^ the strategy applied for enqueuing the output requests
-                                    -> [(a -> Event Bool, Stream p)]
-                                    -- ^ the filters and streams of priorities
-                                    -> Stream a
-                                    -- ^ the input stream
-                                    -> Simulation [Stream a]
-                                    -- ^ the splitted output streams
-splitStreamFilteringPrioritising s fs x =
-  do ref <- liftIO $ newIORef x
-     res <- newResource s 1
-     let stream pred (Cons p) = Cons z where
-           z = do (p', ps) <- p
-                  a <- usingResourceWithPriority res p' $
-                       do p <- liftIO $ readIORef ref
-                          (a, xs) <- runStream p
-                          liftEvent $
-                            do f <- pred a
-                               if f
-                                 then do liftIO $ writeIORef ref xs
-                                         return $ Just a
-                                 else do liftIO $ writeIORef ref $ Cons (return (a, xs))
-                                         return Nothing
-                  case a of
-                    Just a  -> return (a, stream pred ps)
-                    Nothing -> runStream $ stream pred (Cons $ return (p', ps))
-     return $ map (\(pred, ps) -> stream pred ps) fs
 
 -- | Concatenate the input streams applying the 'FCFS' strategy and
 -- producing one output stream.