+++ CHANGELOG.md
@@ -1,8 +1,8 @@
 Version x.x.x
 -----
 
-* Added functions splitStreamFiltering, splitStreamFilteringQueueing and splitStreamFilteringPrioritising
-  to filter when splitting the input stream.
+* Added functions splitStreamFiltering, splitStreamFilteringQueueing to filter when 
+  splitting the input stream.
 
 Version 4.3.1
 -----